server/requirements.txt
# Dépendances de base
numpy>=1.19.0
opencv-python>=4.5.0
Pillow>=8.0.0
scipy>=1.6.0

# PyTorch et dépendances
torch>=1.8.0
torchvision>=0.9.0

# Hugging Face pour le téléchargement des modèles
huggingface-hub>=0.0.12
gdown>=4.4.0

# Utilitaires
tqdm>=4.60.0
python-dotenv>=0.19.0

# Pour le traitement d'image
scikit-image>=0.18.0
